requirements.txt
asgiref==3.2.10
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.1
django-admin-filters==0.1.3
django-autocomplete-light==3.9.7
django-filter==23.5
django-import-export==3.3.4
django-jazzmin==2.6.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
flake8==6.1.0
MarkupPy==1.14
mccabe==0.7.0
odfpy==1.4.1
openpyxl==3.1.2
pycodestyle==2.11.1
pyflakes==3.1.0
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
xlrd==2.0.1
xlwt==1.3.0
